refactor(runtime): move all const enums to flags.ts
