+++ packages/runtime/src/aurelia.ts
@@ -1,6 +1,6 @@
 import { DI, IContainer, IRegistry, PLATFORM, Profiler, Registration } from '@aurelia/kernel';
 import { IDOM, INode } from './dom';
-import { LifecycleFlags } from './observation';
+import { LifecycleFlags } from './flags';
 import { ProxyObserver } from './observation/proxy-observer';
 import { ExposedContext, IRenderingEngine } from './rendering-engine';
 import { CustomElementResource, ICustomElement, ICustomElementType, IProjectorLocator } from './resources/custom-element';
