+++ packages/runtime/src/binding/call.ts
@@ -1,6 +1,7 @@
 import { IServiceLocator, Tracer } from '@aurelia/kernel';
-import { IBindScope, State } from '../lifecycle';
-import { IAccessor, IBindingContext, IObservable, IScope, LifecycleFlags } from '../observation';
+import { LifecycleFlags, State } from '../flags';
+import { IBindScope } from '../lifecycle';
+import { IAccessor, IBindingContext, IObservable, IScope } from '../observation';
 import { IObserverLocator } from '../observation/observer-locator';
 import { hasBind, hasUnbind, IsBindingBehavior } from './ast';
 import { IConnectableBinding } from './connectable';
