+++ packages/runtime/src/binding/ref.ts
@@ -1,6 +1,7 @@
 import { IIndexable, IServiceLocator, Tracer } from '@aurelia/kernel';
-import { IBindScope, State } from '../lifecycle';
-import { IObservable, IScope, LifecycleFlags } from '../observation';
+import { LifecycleFlags, State } from '../flags';
+import { IBindScope } from '../lifecycle';
+import { IObservable, IScope } from '../observation';
 import { hasBind, hasUnbind, IsBindingBehavior } from './ast';
 import { IBinding } from './binding';
 import { IConnectableBinding } from './connectable';
