+++ packages/runtime/src/observation.ts
@@ -1,73 +1,6 @@
 import { IDisposable, IIndexable } from '@aurelia/kernel';
+import { LifecycleFlags } from './flags';
 import { ILifecycle } from './lifecycle';
-
-export enum LifecycleFlags {
-  none                      = 0b00_0000_00000000000000_000_00,
-  mustEvaluate              = 0b00_0001_00000000000000_000_00,
-  mutation                  = 0b00_0000_00000000000000_000_11,
-  isCollectionMutation      = 0b00_0000_00000000000000_000_01,
-  isInstanceMutation        = 0b00_0000_00000000000000_000_10,
-  update                    = 0b00_0000_00000000000000_111_00,
-  updateTargetObserver      = 0b00_0000_00000000000000_001_00,
-  updateTargetInstance      = 0b00_0000_00000000000000_010_00,
-  updateSourceExpression    = 0b00_0000_00000000000000_100_00,
-  from                      = 0b00_0000_11111111111111_000_00,
-  fromFlush                 = 0b00_0000_00000000000111_000_00,
-  fromAsyncFlush            = 0b00_0000_00000000000001_000_00,
-  fromSyncFlush             = 0b00_0000_00000000000010_000_00,
-  fromTick                  = 0b00_0000_00000000000100_000_00,
-  fromStartTask             = 0b00_0000_00000000001000_000_00,
-  fromStopTask              = 0b00_0000_00000000010000_000_00,
-  fromBind                  = 0b00_0000_00000000100000_000_00,
-  fromUnbind                = 0b00_0000_00000001000000_000_00,
-  fromAttach                = 0b00_0000_00000010000000_000_00,
-  fromDetach                = 0b00_0000_00000100000000_000_00,
-  fromCache                 = 0b00_0000_00001000000000_000_00,
-  fromDOMEvent              = 0b00_0000_00010000000000_000_00,
-  fromObserverSetter        = 0b00_0000_00100000000000_000_00,
-  fromBindableHandler       = 0b00_0000_01000000000000_000_00,
-  fromLifecycleTask         = 0b00_0000_10000000000000_000_00,
-  parentUnmountQueued       = 0b00_0010_00000000000000_000_00,
-  // this flag is for the synchronous flush before detach (no point in updating the
-  // DOM if it's about to be detached)
-  doNotUpdateDOM            = 0b00_0100_00000000000000_000_00,
-  isTraversingParentScope   = 0b00_1000_00000000000000_000_00,
-  // Bitmask for flags that need to be stored on a binding during $bind for mutation
-  // callbacks outside of $bind
-  persistentBindingFlags    = 0b11_0000_00000000000000_000_00,
-  allowParentScopeTraversal = 0b01_0000_00000000000000_000_00,
-  useProxies                = 0b10_0000_00000000000000_000_00,
-}
-
-export function stringifyLifecycleFlags(flags: LifecycleFlags): string {
-  const flagNames: string[] = [];
-
-  if (flags & LifecycleFlags.mustEvaluate) { flagNames.push('mustEvaluate'); }
-  if (flags & LifecycleFlags.isCollectionMutation) { flagNames.push('isCollectionMutation'); }
-  if (flags & LifecycleFlags.isInstanceMutation) { flagNames.push('isInstanceMutation'); }
-  if (flags & LifecycleFlags.updateTargetObserver) { flagNames.push('updateTargetObserver'); }
-  if (flags & LifecycleFlags.updateTargetInstance) { flagNames.push('updateTargetInstance'); }
-  if (flags & LifecycleFlags.updateSourceExpression) { flagNames.push('updateSourceExpression'); }
-  if (flags & LifecycleFlags.fromAsyncFlush) { flagNames.push('fromAsyncFlush'); }
-  if (flags & LifecycleFlags.fromSyncFlush) { flagNames.push('fromSyncFlush'); }
-  if (flags & LifecycleFlags.fromStartTask) { flagNames.push('fromStartTask'); }
-  if (flags & LifecycleFlags.fromStopTask) { flagNames.push('fromStopTask'); }
-  if (flags & LifecycleFlags.fromBind) { flagNames.push('fromBind'); }
-  if (flags & LifecycleFlags.fromUnbind) { flagNames.push('fromUnbind'); }
-  if (flags & LifecycleFlags.fromAttach) { flagNames.push('fromAttach'); }
-  if (flags & LifecycleFlags.fromDetach) { flagNames.push('fromDetach'); }
-  if (flags & LifecycleFlags.fromCache) { flagNames.push('fromCache'); }
-  if (flags & LifecycleFlags.fromDOMEvent) { flagNames.push('fromDOMEvent'); }
-  if (flags & LifecycleFlags.fromObserverSetter) { flagNames.push('fromObserverSetter'); }
-  if (flags & LifecycleFlags.fromBindableHandler) { flagNames.push('fromBindableHandler'); }
-  if (flags & LifecycleFlags.fromLifecycleTask) { flagNames.push('fromLifecycleTask'); }
-  if (flags & LifecycleFlags.parentUnmountQueued) { flagNames.push('parentUnmountQueued'); }
-  if (flags & LifecycleFlags.doNotUpdateDOM) { flagNames.push('doNotUpdateDOM'); }
-  if (flags & LifecycleFlags.isTraversingParentScope) { flagNames.push('isTraversingParentScope'); }
-  if (flags & LifecycleFlags.allowParentScopeTraversal) { flagNames.push('allowParentScopeTraversal'); }
-
-  return flagNames.join('|');
-}
 
 export interface IProxyObserver<TObj extends object = object, TMut extends MutationKind = MutationKind.proxy> extends ISubscriberCollection<TMut> {
   proxy: IProxy<TObj>;
