+++ packages/runtime/src/observation/set-observer.ts
@@ -1,6 +1,7 @@
 import { Tracer } from '@aurelia/kernel';
+import { LifecycleFlags } from '../flags';
 import { ILifecycle } from '../lifecycle';
-import { CollectionKind, ICollectionObserver, IObservedSet, LifecycleFlags } from '../observation';
+import { CollectionKind, ICollectionObserver, IObservedSet } from '../observation';
 import { collectionObserver } from './collection-observer';
 
 const proto = Set.prototype;
