+++ packages/runtime/src/resources/binding-behavior.ts
@@ -9,7 +9,8 @@ import {
   Writable
 } from '@aurelia/kernel';
 import { IBinding } from '../binding/binding';
-import { IScope, LifecycleFlags } from '../observation';
+import { LifecycleFlags } from '../flags';
+import { IScope } from '../observation';
 
 export interface IBindingBehavior {
   bind(flags: LifecycleFlags, scope: IScope, binding: IBinding): void;
