+++ packages/runtime/src/resources/binding-behaviors/debounce.ts
@@ -1,7 +1,7 @@
 import { IRegistry, PLATFORM } from '@aurelia/kernel';
 import { Binding, IBinding } from '../../binding/binding';
-import { BindingMode } from '../../binding/binding-mode';
-import { IScope, LifecycleFlags } from '../../observation';
+import { BindingMode, LifecycleFlags } from '../../flags';
+import { IScope } from '../../observation';
 import { BindingBehaviorResource } from '../binding-behavior';
 
 export type DebounceableBinding = IBinding & {
